requirements.txt
langgraph
langchain
langchain_google_genai
langchain_community
pydantic
duckduckgo-search
twilio
Flask
wikipedia
fastapi[standard]
flask_cors
python-dotenv